Split compare.py into 2 scripts
 - compare.py - prints the diff of the measures read from 2 files
 - compare_run.py - does the same thing as compare.py did before

The former will be useful to compare the results from measures taken
using a different "runner". E.g. by running a spec file directly against
an existing cluster.
